refactor : domain 패키지 entity 으로 변경 - domain 패키지 이름이 겹치지 때문에 따로 entity 로 변경했다.
