+++ backend/src/main/java/com/ssafy/home/domain/user/entity/MemberSecret.java
@@ -1,4 +1,4 @@
-package com.ssafy.home.domain.user.domain;
+package com.ssafy.home.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
