+++ backend/src/main/java/com/ssafy/home/domain/user/repository/LoginAttemptCount.java
@@ -1,6 +1,6 @@
 package com.ssafy.home.domain.user.repository;
 
-import com.ssafy.home.domain.user.domain.MemberLoginAttempt;
+import com.ssafy.home.domain.user.entity.MemberLoginAttempt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LoginAttemptCount extends JpaRepository<MemberLoginAttempt, Long> {
