+++ backend/src/main/java/com/ssafy/home/domain/user/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.ssafy.home.domain.user.repository;
 
-import com.ssafy.home.domain.user.domain.Member;
+import com.ssafy.home.domain.user.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
