+++ backend/src/main/java/com/ssafy/home/domain/user/repository/MemberSecurityRepository.java
@@ -1,6 +1,6 @@
 package com.ssafy.home.domain.user.repository;
 
-import com.ssafy.home.domain.user.domain.MemberSecret;
+import com.ssafy.home.domain.user.entity.MemberSecret;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberSecurityRepository extends JpaRepository<MemberSecret, Long> {
